Updates Select component to support option objects
Changes the Select component to accept an array of option objects, where each object has a `label` and a `value` property, instead of a simple array of strings.
This provides more flexibility and allows for displaying different labels than values.
Also adds a `disabled` prop to allow disabling the select component.